cloud_function/requirements.txt
functions-framework
google-cloud-firestore
google-auth>=2.23.0
requests
